requirements.txt
Flask==3.0.3
psycopg[binary]==3.3.3
SQLAlchemy==2.0.36
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
gunicorn==22.0.0
